Tidy up classes for fdl:validations:test
Add some documentation and seperate Compare into its own class
